decode unsigned int fields to long

+++ src/main/scala/Flow.scala
@@ -1,3 +1,3 @@
 
 
-case class Flow(srcIp: String, dstIp: String, packetCount: Int, byteCount: Int, srcPort: Int, dstPort: Int)
+case class Flow(srcIp: String, dstIp: String, packetCount: Long, byteCount: Long, srcPort: Int, dstPort: Int)
